feat: improve memory performance for cached skills

+++ src/skill_loader.rs
@@ -62,11 +62,11 @@ pub struct LoadedSkill {
     pub digest: Digest,
     /// How big the skill was in the registry. Used as a proxy for roughly how large a skill was
     /// and can be used downstream for things like cache eviction.
-    pub size_loaded_from_registry: usize,
+    pub size_loaded_from_registry: u64,
 }
 
 impl LoadedSkill {
-    pub fn new(skill: Arc<dyn Skill>, digest: Digest, size_loaded_from_registry: usize) -> Self {
+    pub fn new(skill: Arc<dyn Skill>, digest: Digest, size_loaded_from_registry: u64) -> Self {
         Self {
             skill,
             digest,
@@ -248,7 +248,7 @@ impl SkillLoaderActor {
         let skill_bytes = registry.load_skill(&skill.name, &skill.tag).await?;
         let SkillImage { bytes, digest } =
             skill_bytes.ok_or_else(|| SkillFetchError::SkillNotFound(skill.clone()))?;
-        let size_loaded_from_registry = bytes.len();
+        let size_loaded_from_registry = bytes.len() as u64;
         let skill = spawn_blocking(move || load_skill_from_wasm_bytes(engine.as_ref(), bytes))
             .await
             .expect("Spawned linking thread must run to completion without being poisoned.")?;
